Independent module for loggers callbacks
