+++ sklearn_genetic/callbacks/__init__.py
@@ -2,7 +2,7 @@ from .early_stoppers import (
     DeltaThreshold,
     ThresholdStopping,
     ConsecutiveStopping,
-    LogbookSaver,
 )
+from .loggers import LogbookSaver
 
-__all__ = ["DeltaThreshold", "ThresholdStopping", "ConsecutiveStopping", "LogbookSaver"]
+__all__ = ["DeltaThreshold", "ThresholdStopping", "ConsecutiveStopping"]
